+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/ClusterFixtureSpec.scala
@@ -1,9 +1,8 @@
-package org.broadinstitute.dsde.workbench.leonardo.cluster
+package org.broadinstitute.dsde.workbench.leonardo
 
 import com.typesafe.scalalogging.LazyLogging
 import org.broadinstitute.dsde.workbench.auth.AuthToken
 import org.broadinstitute.dsde.workbench.fixture._
-import org.broadinstitute.dsde.workbench.leonardo.{Cluster, ClusterRequest, LeonardoTestUtils, MachineConfig}
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 import org.broadinstitute.dsde.workbench.service.Orchestration
 import org.broadinstitute.dsde.workbench.service.test.RandomUtil
